+++ StockSystem/src/main/java/com/cooba/TradeSimulator/Mapper/StockInfoMapper.java
@@ -61,7 +61,7 @@ public interface StockInfoMapper {
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
     @SelectProvider(type=SqlProviderAdapter.class, method="select")
     @Results(id="StockInfoResult", value = {
-        @Result(column="id", property="id", jdbcType=JdbcType.BIGINT, id=true),
+        @Result(column="id", property="id", jdbcType=JdbcType.INTEGER, id=true),
         @Result(column="stockcode", property="stockcode", jdbcType=JdbcType.VARCHAR),
         @Result(column="name", property="name", jdbcType=JdbcType.VARCHAR),
         @Result(column="ISIN_code", property="isinCode", jdbcType=JdbcType.VARCHAR),
@@ -87,7 +87,7 @@ public interface StockInfoMapper {
     }
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
-    default int deleteByPrimaryKey(Long id_) {
+    default int deleteByPrimaryKey(Integer id_) {
         return delete(c -> 
             c.where(id, isEqualTo(id_))
         );
@@ -151,7 +151,7 @@ public interface StockInfoMapper {
     }
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
-    default Optional<StockInfo> selectByPrimaryKey(Long id_) {
+    default Optional<StockInfo> selectByPrimaryKey(Integer id_) {
         return selectOne(c ->
             c.where(id, isEqualTo(id_))
         );
